I added some more tests to cover the different recurrence rules.
