+++ jest.config.js
@@ -2,7 +2,4 @@
 module.exports = {
 	preset: "ts-jest",
 	testEnvironment: "node",
-	transform: {
-		"^.+\\.tsx?$": "ts-jest"
-	}
 };
