+++ package.json
@@ -31,7 +31,6 @@
         "@types/node": "^16.11.6",
         "@types/react": "^17.0.38",
         "@types/react-dom": "^17.0.11",
-        "@types/react-test-renderer": "^18.0.1",
         "@typescript-eslint/eslint-plugin": "^5.2.0",
         "@typescript-eslint/parser": "^5.2.0",
         "builtin-modules": "^3.2.0",
@@ -41,7 +40,6 @@
         "jest": "^29.3.1",
         "obsidian": "^0.16.3",
         "prettier": "^2.5.1",
-        "react-test-renderer": "^17.0.2",
         "ts-jest": "^29.0.3",
         "tslib": "2.3.1",
         "typescript": "4.4.4",
